Parse and include test and exec depends.
Test depends are not used by pacman directly, but Scaled's project support for
pacman projects will use them to handle compiling and running test code.

Exec depends are a way for a package to say "I need this package to be
installed because I'm going to use Scaled's support for forking a JVM using
code from another package to run code in this package". So beyond ensuring that
they are installed when a package is installed, spam does nothing with them.