Reexec installed spam from downloaded spam.
This allows us to have spam update spam. We also urge the user to replace their
downloaded spam with a symlink to the installed version once it exists.

Maybe we should just do this for them, but deleting a shell script out from
under yourself and replacing it with a symlink seems dangerous at best.