Fixed classpath fiddly business.
Pacman (and mfetcher) used to be in the default classloader (because the old
shell script ran java with them in -classpath) but now the default classloader
just contains the bootstrap class, and Pacman and mfetcher are in a child
classloader thereof. So we need that child classloader to be the parent of all
module classloaders rather than the default classloader (which, naturally, is
the default).