A bunch of conflicting depend fiddling.
In the case of system (nee shared) depends, we have to check the stable
(unversioned) id of the depend and avoid allowing multiple versions of the
system depend into the dependency graph.

The resolution process we use is "deepest wins" because any alternative would
require that we know in advance what all possible system depends conflicts were
and resolved accordingly. Since we're trying to operate locally (what are the
depends for this package and its sub-packages), we can't really take a global
view.

This approach is still be prone to failure if two transitive dependencies
depend on different versions of a system depend (i.e. module C depends on
module B and module A; module B depends on sysdep 1.0 and module A depends on
sysdep 1.1).

However, that's so far not a problem because the only actual sysdep we have
(other than JDK 1.8) is scala-library and everything that uses Scaled depends
on scaled-api which exports the "canonical" version of scala-library for use by
every other module. So all other modules will filter their own scala-library
depend (if they happen to inherit it from some Maven artifact, say, and no one
other than scaled#api will ever manage to export a scala-library depend.

So it's still a fragile hack, but less fragile than just assuming everything,
including all random third-party Scala libraries, will depend on the right
Scala version. We still have an implicit dependency on a particular Scala major
version (at present 2.11), and any module that attempts to depend on a library
built for a different Scala major version will find itself crapping its pants
in short order. But that should be immediately apparent to the module
developer, so such things should never get out into the wild. But we shouldn't
burden the module developer with caring about Scala minor versions because
that's not something they'll have control over.