Fix some ordering issues in upgrade rebuilds.
When building the packages on which the target upgrade package depends, we immediately ran off and
rebuilt all the other packages that depend on those packages, which was potentially problematic
because those packages might depend on the package we're trying to upgrade in the first place.

Now we just rebuild/upgrade the packages on which our target package depends, then rebuild our
target package, then go back and rebuild any packages which depend on the packages that got
upgraded, so as to ensure that we never rebuild a package without first rebuilding its
dependencies.

It's all a bit messy and I should really reorg all of this and handle it in a more disciplined way,
but we'll see how long this bandaid lasts for now.